quick edit

+++ WeirdAlgorithm.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class WeirdAlgorithm {
     public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int n = sc.nextInt();
+	    long n = sc.nextLong();
 		
             System.out.print(n + " ");
